[ UPDATE ] slider가 이동할 때 indicator도 함께 작동되게 하라 - indicator를 클릭하면 해당 slider의 indicator의 스타일을 변경하였다 - 해당 슬라이더로 이동하게 구현하였다. - useEffect를 사용하여 화면이 렌더링 됨과 동시에 자동으로 autoPlay 되도록 설정하였다. - 이는 pause/play버튼을 누르면 useEffect가 업데이트 되도록 의존성배열에 연결하였다
