--- image-slider-react/src/style.css
@@ -1,18 +1,0 @@
-.indicator {
-  border: 1px solid red;
-}
-
-.indicator ul {
-  display: flex;
-  justify-content: center;
-}
-
-.indicator ul li {
-  width: 15px;
-  height: 15px;
-  margin: 0 5px;
-  list-style: none;
-  cursor: pointer;
-  border-radius: 50%;
-  background-color: teal;
-}
